+++ src/app-toolkit/helpers/position/vault.token-helper.ts
@@ -1,13 +1,15 @@
+import { isPromise } from 'util/types';
+
 import { Inject, Injectable } from '@nestjs/common';
-import { BigNumberish } from 'ethers';
 import { compact } from 'lodash';
 
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
 import {
   buildDollarDisplayItem,
+  buildNumberDisplayItem,
   buildPercentageDisplayItem,
 } from '~app-toolkit/helpers/presentation/display-item.present';
-import { getImagesFromToken, getLabelFromToken } from '~app-toolkit/helpers/presentation/image.present';
+import { getImagesFromToken } from '~app-toolkit/helpers/presentation/image.present';
 import { EthersMulticall } from '~multicall';
 import { ContractType } from '~position/contract.interface';
 import { StatsItem } from '~position/display.interface';
@@ -15,53 +17,56 @@ import { AppTokenPosition, Token } from '~position/position.interface';
 import { AppGroupsDefinition } from '~position/position.service';
 import { Network } from '~types/network.interface';
 
-import { YearnContractFactory, YearnVault } from '../contracts';
-
-type YearnVaultTokenDataProps = {
+export type VaultTokenDataProps = {
   liquidity: number;
   reserve: number;
 };
 
-export type YearnVaultTokenHelperParams<T> = {
+export type VaultTokenHelperParams<T> = {
   network: Network;
   appId: string;
   groupId: string;
   dependencies?: AppGroupsDefinition[];
-  resolvePrimaryLabel?: (opts: { symbol: string; vaultAddress: string; underlyingToken: Token }) => string;
-  resolveVaultAddresses: (opts: { multicall: EthersMulticall; network: Network }) => Promise<string[]>;
-  resolveContract?: (opts: { address: string; network: Network }) => T;
-  resolveUnderlyingToken?: (opts: { multicall: EthersMulticall; contract: T }) => Promise<string | null>;
-  resolvePricePerShare?: (opts: { multicall: EthersMulticall; contract: T }) => Promise<BigNumberish>;
-  resolvePricePerShareActual?: (opts: { pricePerShareRaw: BigNumberish; decimals: number }) => number;
+  resolveContract: (opts: { address: string; network: Network }) => T;
+  resolveVaultAddresses: (opts: { multicall: EthersMulticall; network: Network }) => string[] | Promise<string[]>;
+  resolveUnderlyingTokenAddress: (opts: { multicall: EthersMulticall; contract: T }) => string | Promise<string | null>;
+  resolveReserve: (opts: {
+    address: string;
+    contract: T;
+    multicall: EthersMulticall;
+    underlyingToken: Token;
+    network: Network;
+  }) => number | Promise<number>;
+  resolvePricePerShare: (opts: {
+    contract: T;
+    multicall: EthersMulticall;
+    underlyingToken: Token;
+    reserve: number;
+    supply: number;
+  }) => number | Promise<number>;
   resolveApy?: (opts: { vaultAddress: string; multicall: EthersMulticall; contract: T }) => Promise<number>;
+  resolvePrimaryLabel?: (opts: { symbol: string; vaultAddress: string; underlyingToken: Token }) => string;
+  resolveImages?: (opts: { underlyingToken: Token }) => string[];
 };
 
 @Injectable()
-export class YearnLikeVaultTokenHelper {
-  constructor(
-    @Inject(APP_TOOLKIT) private readonly appToolkit: IAppToolkit,
-    @Inject(YearnContractFactory) private readonly yearnContractFactory: YearnContractFactory,
-  ) {}
+export class VaultTokenHelper {
+  constructor(@Inject(APP_TOOLKIT) private readonly appToolkit: IAppToolkit) {}
 
-  async getTokens<T = YearnVault>({
+  async getTokens<T>({
     network,
     appId,
     groupId,
     dependencies = [],
     resolveVaultAddresses,
-    resolveContract = ({ address, network }) =>
-      this.yearnContractFactory.yearnVault({ address, network }) as unknown as T,
-    resolvePrimaryLabel = ({ underlyingToken }) => `${getLabelFromToken(underlyingToken)} Vault`,
-    resolveUnderlyingToken = ({ contract, multicall }) =>
-      multicall
-        .wrap(contract as unknown as YearnVault)
-        .token()
-        .catch(() => null),
-    resolvePricePerShare = ({ contract, multicall }) =>
-      multicall.wrap(contract as unknown as YearnVault).getPricePerFullShare(),
-    resolvePricePerShareActual = ({ pricePerShareRaw, decimals }) => Number(pricePerShareRaw) / 10 ** decimals,
+    resolveContract,
+    resolveUnderlyingTokenAddress,
+    resolveReserve,
+    resolvePricePerShare,
     resolveApy,
-  }: YearnVaultTokenHelperParams<T>) {
+    resolvePrimaryLabel = ({ symbol }) => symbol,
+    resolveImages = ({ underlyingToken }) => getImagesFromToken(underlyingToken),
+  }: VaultTokenHelperParams<T>) {
     const multicall = this.appToolkit.getMulticall(network);
 
     const baseTokens = await this.appToolkit.getBaseTokenPrices(network);
@@ -72,29 +77,31 @@ export class YearnLikeVaultTokenHelper {
     const curvePoolTokens = await Promise.all(
       vaultAddresses.map(async vaultAddress => {
         const type = ContractType.APP_TOKEN;
-        const erc20Contract = this.yearnContractFactory.erc20({ address: vaultAddress, network });
-        const vaultContract = resolveContract({ address: vaultAddress, network });
+        const erc20Contract = this.appToolkit.globalContracts.erc20({ address: vaultAddress, network });
+        const contract = resolveContract({ address: vaultAddress, network });
 
-        const [symbol, decimals, supplyRaw, underlyingTokenAddressRaw, pricePerShareRaw, apy] = await Promise.all([
+        const [symbol, decimals, supplyRaw] = await Promise.all([
           multicall.wrap(erc20Contract).symbol(),
           multicall.wrap(erc20Contract).decimals(),
           multicall.wrap(erc20Contract).totalSupply(),
-          resolveUnderlyingToken({ multicall, contract: vaultContract }),
-          resolvePricePerShare({ multicall, contract: vaultContract }),
-          resolveApy ? resolveApy({ multicall, contract: vaultContract, vaultAddress }) : Promise.resolve(0),
         ]);
 
+        // Find underlying token in dependencies
+        const underlyingTokenAddressRawMaybePromise = resolveUnderlyingTokenAddress({ multicall, contract });
+        const underlyingTokenAddressRaw = isPromise(underlyingTokenAddressRawMaybePromise)
+          ? await underlyingTokenAddressRawMaybePromise.catch(() => null)
+          : underlyingTokenAddressRawMaybePromise;
         if (!underlyingTokenAddressRaw) return null;
 
-        // Find underlying token in dependencies
         const underlyingTokenAddress = underlyingTokenAddressRaw.toLowerCase();
         const underlyingToken = allTokens.find(p => p.address === underlyingTokenAddress);
         if (!underlyingToken) return null;
 
         // Data props
         const supply = Number(supplyRaw) / 10 ** decimals;
-        const pricePerShare = resolvePricePerShareActual({ pricePerShareRaw, decimals });
-        const reserve = pricePerShare * supply;
+        const reserve = await resolveReserve({ address: vaultAddress, contract, multicall, underlyingToken, network });
+        const pricePerShare = await resolvePricePerShare({ multicall, contract, reserve, supply, underlyingToken });
+        const apy = resolveApy ? await resolveApy({ multicall, contract, vaultAddress }) : 0;
         const price = underlyingToken.price * pricePerShare;
         const tokens = [underlyingToken];
         const liquidity = price * supply;
@@ -103,13 +110,14 @@ export class YearnLikeVaultTokenHelper {
         const label = resolvePrimaryLabel({ vaultAddress, symbol, underlyingToken });
         const secondaryLabel = buildDollarDisplayItem(price);
         const tertiaryLabel = resolveApy ? `${(apy * 100).toFixed(3)}% APY` : '';
-        const images = getImagesFromToken(underlyingToken);
-        const statsItems: StatsItem[] = [];
+        const images = resolveImages({ underlyingToken });
+        const statsItems: StatsItem[] = [{ label: 'Liquidity', value: buildDollarDisplayItem(liquidity) }];
+        if (reserve > 0) statsItems.push({ label: 'Reserve', value: buildNumberDisplayItem(reserve) });
         if (resolveApy) statsItems.push({ label: 'APY', value: buildPercentageDisplayItem(apy) });
 
-        const vaultToken: AppTokenPosition<YearnVaultTokenDataProps> = {
-          address: vaultAddress,
+        const vaultToken: AppTokenPosition<VaultTokenDataProps> = {
           type,
+          address: vaultAddress,
           network,
           appId,
           groupId,