+++ src/apps/balancer-v2/balancer-v2.module.ts
@@ -4,6 +4,7 @@ import { CurveAppModule } from '~apps/curve';
 
 import { ArbitrumBalancerV2BalanceFetcher } from './arbitrum/balancer-v2.balance-fetcher';
 import { ArbitrumBalancerV2PoolTokenFetcher } from './arbitrum/balancer-v2.pool.token-fetcher';
+import { ArbitrumBalancerV2StakedfContractPositionFetcher } from './arbitrum/balancer-v2.staked.contract-position-fetcher';
 import BALANCER_V2_DEFINITION, { BalancerV2AppDefinition } from './balancer-v2.definition';
 import { BalancerV2ContractFactory } from './contracts';
 import { EthereumBalancerV2BalanceFetcher } from './ethereum/balancer-v2.balance-fetcher';
@@ -15,10 +16,12 @@ import { BalancerV2ClaimableContractPositionBalanceHelper } from './helpers/bala
 import { BalancerV2EventsPoolTokenDataStrategy } from './helpers/balancer-v2.events.pool-token-address-strategy';
 import { BalancerV2GaugeAddressesGetter } from './helpers/balancer-v2.gauge-addresses-getter';
 import { BalancerV2PoolTokensHelper } from './helpers/balancer-v2.pool.token-helper';
+import { BalancerV2GaugeRewardTokenStrategy } from './helpers/balancer-v2.reward-token-strategy';
 import { BalancerV2SpotPriceHelper } from './helpers/balancer-v2.spot-price.helper';
 import { BalancerV2TheGraphPoolTokenDataStrategy } from './helpers/balancer-v2.the-graph.pool-token-address-strategy';
 import { PolygonBalancerV2BalanceFetcher } from './polygon/balancer-v2.balance-fetcher';
 import { PolygonBalancerV2PoolTokenFetcher } from './polygon/balancer-v2.pool.token-fetcher';
+import { PolygonBalancerV2StakedfContractPositionFetcher } from './polygon/balancer-v2.staked.contract-position-fetcher';
 
 @Register.AppModule({
   appId: BALANCER_V2_DEFINITION.id,
@@ -29,6 +32,7 @@ import { PolygonBalancerV2PoolTokenFetcher } from './polygon/balancer-v2.pool.to
     // Arbitrum
     ArbitrumBalancerV2BalanceFetcher,
     ArbitrumBalancerV2PoolTokenFetcher,
+    ArbitrumBalancerV2StakedfContractPositionFetcher,
     // Ethereum
     EthereumBalancerV2BalanceFetcher,
     EthereumBalancerV2PoolTokenFetcher,
@@ -37,6 +41,7 @@ import { PolygonBalancerV2PoolTokenFetcher } from './polygon/balancer-v2.pool.to
     // Polygon
     PolygonBalancerV2BalanceFetcher,
     PolygonBalancerV2PoolTokenFetcher,
+    PolygonBalancerV2StakedfContractPositionFetcher,
     // Helpers
     BalancerV2CacheManager,
     BalancerV2PoolTokensHelper,
@@ -45,6 +50,7 @@ import { PolygonBalancerV2PoolTokenFetcher } from './polygon/balancer-v2.pool.to
     BalancerV2ClaimableContractPositionBalanceHelper,
     BalancerV2EventsPoolTokenDataStrategy,
     BalancerV2GaugeAddressesGetter,
+    BalancerV2GaugeRewardTokenStrategy,
   ],
   exports: [BalancerV2SpotPriceHelper, BalancerV2PoolTokensHelper, BalancerV2ContractFactory],
 })