+++ src/apps/balancer-v2/ethereum/balancer-v2.staked.contract-position-fetcher.ts
@@ -26,7 +26,6 @@ export class EthereumBalancerV2StakedfContractPositionFetcher implements Positio
   ) {}
 
   async getPositions() {
-    const network = Network.ETHEREUM_MAINNET;
     const farms = await this.balancerV2GaugeAddressesGetter.getGauges({ network });
     return await this.curveStakingHelper.getContractPositions<BalancerGauge>({
       network,