+++ src/apps/olympus/ethereum/olympus.g-ohm.token-fetcher.ts
@@ -27,14 +27,14 @@ export class EthereumOlympusGOhmTokenFetcher implements PositionFetcher<AppToken
   ) {}
 
   async getPositions() {
-    return this.appToolkit.helpers.singleVaultTokenHelper.getTokens<OlympusGOhmToken>({
+    return this.appToolkit.helpers.vaultTokenHelper.getTokens<OlympusGOhmToken>({
       appId: OLYMPUS_DEFINITION.id,
       groupId: OLYMPUS_DEFINITION.groups.gOhm.id,
       network: Network.ETHEREUM_MAINNET,
       dependencies: [{ appId, groupIds: [OLYMPUS_DEFINITION.groups.sOhm.id], network }],
-      address: '0x0ab87046fbb341d058f17cbc4c1133f25a20a52f', // gOHM
+      resolveVaultAddresses: () => ['0x0ab87046fbb341d058f17cbc4c1133f25a20a52f'], // gOHM
       resolveContract: ({ address, network }) => this.contractFactory.olympusGOhmToken({ address, network }),
-      resolveUnderlyingTokenAddress: () => '0x04906695d6d12cf5459975d7c3c03356e4ccd460', // sOHM
+      resolveUnderlyingTokenAddress: ({ contract, multicall }) => multicall.wrap(contract).sOHM(),
       resolvePricePerShare: async ({ multicall, contract, underlyingToken }) => {
         const oneOhm = BigNumber.from(1).mul(10).pow(underlyingToken.decimals);
         const [gOhmDecimalsRaw, gOhmConvertedAmountRaw] = await Promise.all([