+++ src/apps/synthetix/synthetix.definition.ts
@@ -8,9 +8,9 @@ export const SYNTHETIX_DEFINITION = appDefinition({
   name: 'Synthetix',
   description: `A new financial primitive enabling the creation of synthetic assets, offering unique derivatives and exposure to real-world assets on the blockchain.`,
   groups: {
-    synth: { id: 'synth', type: GroupType.TOKEN },
-    farm: { id: 'farm', type: GroupType.POSITION },
-    mintr: { id: 'mintr', type: GroupType.POSITION },
+    synth: { id: 'synth', type: GroupType.TOKEN, label: 'Synths' },
+    farm: { id: 'farm', type: GroupType.POSITION, label: 'Mintr' },
+    mintr: { id: 'mintr', type: GroupType.POSITION, label: 'Staking' },
   },
   url: 'https://synthetix.io/',
   links: {