+++ src/apps/tokemak/tokemak.definition.ts
@@ -8,8 +8,8 @@ export const TOKEMAK_DEFINITION = appDefinition({
   name: 'Tokemak',
   description: `Tokemak creates sustainable DeFi liquidity and capital efficient markets through a convenient decentralized market making protocol.`,
   groups: {
-    farm: { id: 'farm', type: GroupType.POSITION },
-    reactor: { id: 'reactor', type: GroupType.TOKEN },
+    farm: { id: 'farm', type: GroupType.POSITION, label: 'Staking' },
+    reactor: { id: 'reactor', type: GroupType.TOKEN, label: 'Reactors' },
   },
   url: 'https://www.tokemak.xyz/',
   links: {