+++ src/apps/trader-joe/contracts/ethers/index.ts
@@ -8,6 +8,7 @@ export type { TraderJoeChefV3 } from './TraderJoeChefV3';
 export type { TraderJoeChefV3Rewarder } from './TraderJoeChefV3Rewarder';
 export type { TraderJoeStableStaking } from './TraderJoeStableStaking';
 export type { TraderJoeVeJoeStaking } from './TraderJoeVeJoeStaking';
+export type { TraderJoeXJoe } from './TraderJoeXJoe';
 export * as factories from './factories';
 export { TraderJoeChefBoosted__factory } from './factories/TraderJoeChefBoosted__factory';
 export { TraderJoeChefV2Rewarder__factory } from './factories/TraderJoeChefV2Rewarder__factory';
@@ -16,3 +17,4 @@ export { TraderJoeChefV3Rewarder__factory } from './factories/TraderJoeChefV3Rew
 export { TraderJoeChefV3__factory } from './factories/TraderJoeChefV3__factory';
 export { TraderJoeStableStaking__factory } from './factories/TraderJoeStableStaking__factory';
 export { TraderJoeVeJoeStaking__factory } from './factories/TraderJoeVeJoeStaking__factory';
+export { TraderJoeXJoe__factory } from './factories/TraderJoeXJoe__factory';